Splitting code into multiple .py files by classes
