+++ UNO Project/uno_game.py
@@ -1,89 +1,8 @@
 import random
 import time
-
-class Card:
-
-    def __init__(self, color, rank):
-        self.rank = rank
-        self.ctype = {
-            '0':'number','1':'number','2':'number','3':'number','4':'number','5':'number','6':'number',
-            '7':'number','8':'number','9':'number','Skip':'action','Reverse':'action','Draw2':'action',
-            'Draw4':'action_nocolor','Wild':'action_nocolor'
-        }
-
-        if self.ctype[rank] == 'number':
-            self.color = color
-            self.cardtype = 'number'
-        elif self.ctype[rank] == 'action':
-            self.color = color
-            self.cardtype = 'action'
-        else:
-            self.color = None
-            self.cardtype = 'action_nocolor'
-
-    def __str__(self):
-        if self.color == None:
-            return self.rank
-        else:
-            return self.color + " " + self.rank
-        
-class Deck:
-
-    def __init__(self):
-        self.deck = []
-        self.rank = ('0','1','2','3','4','5','6','7','8','9','Skip','Reverse','Draw2','Draw4','Wild')
-        self.color= ('RED','GREEN','BLUE','YELLOW')
-
-        for clr in self.color:
-            for ran in self.rank:
-                # if ctype[ran] != 'action_nocolor':
-                self.deck.append(Card(clr, ran))
-                self.deck.append(Card(clr, ran))
-                # else:
-                #     self.deck.append(Card(clr, ran))
-
-    def __str__(self):
-        deck_comp = ''
-        for card in self.deck:
-            deck_comp += '\n' + card.__str__()
-        return 'The deck has ' + deck_comp
-
-    def shuffle(self):
-        random.shuffle(self.deck)
-
-    def deal(self):
-        return self.deck.pop()
-
-class Hand:
-
-    def __init__(self):
-        self.cards = []
-        self.cardsstr = []
-        self.number_cards = 0
-        self.action_cards = 0
-
-    def add_card(self, card):
-        self.cards.append(card)
-        self.cardsstr.append(str(card))
-        if card.cardtype == 'number':
-            self.number_cards += 1
-        else:
-            self.action_cards += 1
-
-    def remove_card(self, place):
-        self.cardsstr.pop(place - 1)
-        return self.cards.pop(place - 1)
-
-    def cards_in_hand(self):
-        for i in range(len(self.cardsstr)):
-            print(f' {i + 1}.{self.cardsstr[i]}')
-
-    def single_card(self, place):
-        return self.cards[place - 1]
-
-    def no_of_cards(self):
-        return len(self.cards)
-    
+from card import Card
+from deck import Deck
+from hand import Hand   
 
 #Funciton to randomly select who starts first
 def choose_first():
